Move horizontal initiative profile in to the standard profile panel
