--- client/src/panels/panel_declarations/intro/horizontal_initiative_profile.yaml
@@ -1,25 +1,0 @@
-horizontal_initiative_profile:
-  en: Horizontal Initiative Profile
-  fr: Profile d’initiative horizontale
-
-hi_name:
-  en: Horizontal Initiative
-  fr: Initiative horizontale
-hi_lead_dept:
-  en: Lead department
-  fr: Ministère responsable
-hi_start_year:
-  en: Started year
-  fr: Annéé de début
-hi_end_year:
-  en: Ending year
-  fr: Année de fin
-hi_total_allocated_amount:
-  en: Total allocated amount
-  fr: Montant total alloué
-hi_website:
-  en: Website
-  fr: Site web
-hi_dr_link:
-  en: Most recent report
-  fr: Dernier rapport
