+++ client/src/panels/panel_declarations/intro/index.js
@@ -1,5 +1,3 @@
-export { declare_horizontal_initiative_profile_panel } from './horizontal_initiative_profile.js';
-
 export {
   declare_profile_panel,
   declare_description_panel,
